+++ test/BM25Spec.hs
@@ -26,6 +26,7 @@ mkActivity name loc xs =
     Activity
         { activityName = name
         , activityDescription = []
+        , activityDocumentation = []
         , activitySynonyms = M.empty
         , activityClassification = M.empty
         , activityLocation = loc