+++ test/CrossDBActivityLinkSpec.hs
@@ -66,6 +66,7 @@ mkActivity name loc exs =
     Activity
         { activityName = name
         , activityDescription = []
+        , activityDocumentation = []
         , activitySynonyms = M.empty
         , activityClassification = M.empty
         , activityLocation = loc