+++ test/JournalSpec.hs
@@ -421,6 +421,7 @@ supplierActivity =
     Activity
         { activityName = "milk production"
         , activityDescription = []
+        , activityDocumentation = []
         , activitySynonyms = M.empty
         , activityClassification = M.empty
         , activityLocation = "FR"