+++ test/PersistenceSpec.hs
@@ -406,6 +406,7 @@ milkActivity name =
     Activity
         { activityName = name
         , activityDescription = []
+        , activityDocumentation = []
         , activitySynonyms = M.empty
         , activityClassification = M.empty
         , activityLocation = "FR"