+++ test/SetupInfoSpec.hs
@@ -55,6 +55,7 @@ minimalActivity name exs =
     Activity
         { activityName = name
         , activityDescription = []
+        , activityDocumentation = []
         , activitySynonyms = M.empty
         , activityClassification = M.empty
         , activityLocation = "GLO"